chore: moved edit destination into destination folder
